Modify OSContext
