+++ src/main/kotlin/model/os/OSContext.kt
@@ -5,3 +5,23 @@ interface OSContext {
     val settingPath: String
     val desktopPath: String
 }
+
+class OSContextForMac: OSContext {
+    override val type: OSType = OSType.MAC_OS
+    override val settingPath: String =
+        System.getProperty("user.home") + "/Library/Application Support/ScrcpyHub/"
+    override val desktopPath: String =
+        System.getProperty("user.home") + "/Desktop/"
+}
+
+class OSContextForLinux: OSContext {
+    override val type: OSType = OSType.LINUX
+    override val settingPath: String = System.getProperty("user.home") + "/ScrcpyHub/"
+    override val desktopPath: String = System.getProperty("user.home") + "/Desktop/"
+}
+
+class OSContextForWindows: OSContext {
+    override val type: OSType = OSType.WINDOWS
+    override val settingPath: String = "./"
+    override val desktopPath: String = System.getProperty("user.home") + "/Desktop/"
+}
