+++ src/main/kotlin/model/os/OSType.kt
@@ -5,3 +5,11 @@ enum class OSType {
     LINUX,
     WINDOWS,
 }
+
+fun getOSType(): OSType {
+    return when (System.getProperty("os.name")) {
+        "Mac OS X" -> OSType.MAC_OS
+        "Linux" -> OSType.LINUX
+        else -> OSType.WINDOWS
+    }
+}
